% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportCSV}
\alias{exportCSV}
\title{Export a file}
\usage{
exportCSV(x, file, ...)
}
\arguments{
\item{x}{a data frame or tibble to export}

\item{file}{file or connection to write to}

\item{...}{arguments to pass to \code{readr::write_csv()} or \code{xlsx::write.xlsx}}
}
\value{
The file path of the exported file.
}
\description{
Handy wrapper functions to export data tables that then returns the file path of the exported file. This can be useful in applications such as in conjuction with \code{tarchetypes::tar_file}.
}
\details{
If the file path directory does not exist, the directory is created prior to export.
}
\examples{
## Export to a csv
exportCSV(iris, paste0(tempdir(),"/iris.csv"))

## Export to an Excel workbook
exportCSV(iris, paste0(tempdir(),"/iris.xlsx"))
}
