% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomatch.R
\name{nomatch}
\alias{nomatch}
\alias{\%!in\%}
\title{Non-matching operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{vector or NULL: the values to be matched}

\item{y}{vector or NULL: the values to be matched against}
}
\description{
A binary operator to detect non matching values. The opposite of the \code{\%in\%} operator.
}
\examples{
1:10 \%!in\% c(1,3,5,9)
}
