% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{suitableParallelPlan}
\alias{suitableParallelPlan}
\alias{suitableParallelStrategy}
\alias{suitableParallelWorkers}
\title{Generate a suitable parallel processing plan}
\usage{
suitableParallelPlan(
  strategy = suitableParallelStrategy(),
  workers = suitableParallelWorkers(),
  ...
)

suitableParallelStrategy()

suitableParallelWorkers(RAM_per_worker = 8, proportion_max_workers = 0.75)
}
\arguments{
\item{strategy}{The parallel strategy to use. See \code{?future::plan} for details.}

\item{workers}{The number of workers to use.}

\item{...}{Arguments to pass to \code{future::plan()}.}

\item{RAM_per_worker}{The memory (in gigabytes) to allocate to each worker when calculating a suitable
number of workers.}

\item{proportion_max_workers}{A value between 0 and 1. The maximum number of workers to allocate as a
proportion of the total number of workers available (as detected by \code{future::availableCores()}).}
}
\description{
Generate a suitable parallel processing future plan for your session. See \code{?future::plan}
for more details on future plans.
}
\examples{
## Return a suitable parallel strategy for the current session
suitableParallelStrategy()

## Return a suitable number parallel workers based on the system resources.
suitableParallelWorkers()

\dontrun{
## Generate a suitable parallel plan
suitableParallelPlan()
}
}
