% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operators.R
\name{\%+\%}
\alias{\%+\%}
\alias{\%+ \%}
\alias{\%+_\%}
\alias{\%+-\%}
\title{String operators}
\usage{
x \%+\% y

x \%+ \% y

x \%+_\% y

x \%+-\% y
}
\arguments{
\item{x}{a vector coercible to a character string}

\item{y}{a vector coercible to a character string}
}
\value{
A vector of concatenated strings.
}
\description{
String operators to concatenate vectors that are coercible to character strings. Vector arguments are recycled as needed.
}
\examples{
## Concatenate two strings
'fish' \%+\% 'sticks'

## Concatenate a numeric and string
1 \%+\% '1'

## Concatenate two character vectors
letters[1:5] \%+\% letters[1:5]

## Concatenate two strings with white space between
'hello' \%+ \% 'world'

## Concatenate two character vectors with white space between
letters[1:5] \%+ \% letters[1:5]

## Concatenate two strings with an underscore between
'hello' \%+_\% 'world'

## Concatenate two character vectors with an underscore between
letters[1:5] \%+_\% letters[1:5]

## Concatenate two strings with a dash between
'hello' \%+-\% 'world'

## Concatenate two character vectors with a dash between
letters[1:5] \%+-\% letters[1:5]
}
