% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingChecks.R
\name{writingChecks}
\alias{writingChecks}
\title{Writing checks}
\usage{
writingChecks(
  file = "manuscript.Rmd",
  checks = c("word count", "spelling", "grammar")
)
}
\arguments{
\item{file}{Path to a \code{.Rmd} file to check.}

\item{checks}{Character vector of checks to perform.}
}
\value{
None. Check results are printed to the console.
}
\description{
Check an R markdown file (.Rmd) for word count, spelling and grammatical errors.
}
\examples{
 file_path <- system.file('example.Rmd',package = 'jfmisc')
 writingChecks(file_path)
}
