% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_neat.R
\name{theme_neat}
\alias{theme_neat}
\title{A neat ggplot2 theme}
\usage{
theme_neat(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts}

\item{base_family}{base font family}
}
\description{
A ggplot2 theme that includes bold face titles and
neat facet themes.
}
\examples{
iris \%>\% 
  ggplot2::ggplot(ggplot2::aes(x = Sepal.Length,
                               y = Sepal.Width,
                               colour = Species)) +
  ggplot2::geom_point() +
  ggplot2::facet_wrap(~Species) +
  theme_neat() 
}
